+++ API/Problem-API/src/main/java/tw/waterball/judgegirl/problemapi/clients/FakeProblemServiceDriver.java
@@ -1,4 +1,4 @@
-package tw.waterball.judgegirl.studentapi.clients;
+package tw.waterball.judgegirl.problemapi.clients;
 
 import tw.waterball.judgegirl.commons.exceptions.NotFoundException;
 import tw.waterball.judgegirl.commons.models.files.FileResource;
