+++ Spring-Boot/Spring-Boot-Commons/src/main/java/tw/waterball/judgegirl/springboot/configs/ServiceDriverConfiguration.java
@@ -19,10 +19,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tw.waterball.judgegirl.api.retrofit.RetrofitFactory;
 import tw.waterball.judgegirl.commons.token.TokenService;
+import tw.waterball.judgegirl.problemapi.clients.ProblemApiClient;
+import tw.waterball.judgegirl.problemapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.springboot.configs.properties.ServiceProps;
 import tw.waterball.judgegirl.springboot.profiles.productions.ServiceDriver;
-import tw.waterball.judgegirl.studentapi.clients.ProblemApiClient;
-import tw.waterball.judgegirl.studentapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.studentapi.clients.StudentApiClient;
 import tw.waterball.judgegirl.studentapi.clients.StudentServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.BagInterceptor;
